requirements.txt
feedparser>=6.0
trafilatura>=2.0
requests>=2.28
beautifulsoup4>=4.12
openai>=1.0
jinja2>=3.1
python-dotenv>=1.0
pyyaml>=6.0
